feat(tests): initial setup for tests
This is an initial version of automated tests for Homestead, which will
be very useful as the project is getting more complex and it’s hard to
keep track of everything manually.

I don’t fully understand how everything works yet, but it’s now running
on CI in Github Actions. I got a lot of inspiration from this repo:
renatoathaydes/lisp-site-gen.

Also today, in Moscow, Alexei Navalny was buried. He was only 47 years
old. It feels like such a tragic mistake that he is now dead, and the
person who killed him is still president. I watched the live stream of
the religious service and the burial, and it was extremely emotional to
see how many people showed up, despite the risk. How many people changed
“No to war”. How many more people could not be there. There’s a lot of
us, and that makes me feel warm.